New "download list" plugin for the Zenphoto releases
This plugin replaces the function for the codeblock on the
"older-versions-archive" page of the site. It allows for automating the
support release process and provides links to older versions of zenphoto
at their "cloud" storage location. (Either GitHub or GoogleCode
archive.) This allows us to remove old files from our server.