get support and dev build info from articles
Articles must be made in the "release" category to announce the
releases. For "current" releases the articles must be published. The
first published article with no tag (e.g. not BETA nor RCx) is
considered the "current" release for the download button.

The first published release is considered used for the "support" release
for the current version.  E.g. it occupies the first link under "change
log"

The first un-published release is considered the "dev" release. E.g. it
occupies the second link under "change log". If there is no unpublished
release there will be no link.